src/main/java/com/example/passportVerify/passportVerifyBack/service/UserDetailsInfo.java
package com.example.passportVerify.passportVerifyBack.service;

import com.example.passportVerify.passportVerifyBack.entity.User;
import org.springframework.security.core.GrantedAuthority;

import java.util.Collection;
import java.util.List;

public class UserDetailsInfo implements org.springframework.security.core.userdetails.UserDetails {
    private String email;
    private String password;
    private List<GrantedAuthority> authorities;
    public UserDetailsInfo(User user) {
        email = user.getEmail();
        password = user.getPassword();

    }
    @Override
    public Collection<? extends GrantedAuthority> getAuthorities() {
        return authorities;
    }

    @Override
    public String getPassword() {
        return password;
    }

    @Override
    public String getUsername() {
        return email;
    }


    @Override
    public boolean isAccountNonExpired() {
        return true;
    }

    @Override
    public boolean isAccountNonLocked() {
        return true;
    }

    @Override
    public boolean isCredentialsNonExpired() {
        return true;
    }

    @Override
    public boolean isEnabled() {
        return true;
    }
}
